Reorganize tests: test_indicators, test_expressions, test_primitives
- Remove test_pandas.py and test_polars.py
- test_indicators.py: pandas indicators across all freqs + xfail for polars input
- test_expressions.py: polars expressions across all freqs (all expressions)
- test_primitives.py: all primitives across both backends + freqs
- Fix ZigZag: return int dtype arrays to avoid IndexError on empty result
- Fix dataframe_eval: promote Series to DataFrame before eval so string
  expressions can reference the column by name

+++ src/mplchart/primitives/zigzag.py
@@ -66,7 +66,7 @@ def calc_zigzag(prices, threshold=5.0):
                 values.append(pv)
                 pi, pv, pdir = i, h, +1
 
-    return np.array(index), np.array(values)
+    return np.array(index, dtype=int), np.array(values, dtype=float)
 
 
 class ZigZag(Primitive):
